+++ MDAnalysisData/tests/__init__.py
@@ -1,0 +1,3 @@
+# tests/__init__.py so that pytest-cov collects data
+# otherwise "CoverageWarning: No data was collected. (no-data-collected)"
+# see https://stackoverflow.com/a/60579142